- Adicionando arquivos para a entidade User e arrumando caracteres errados de mensagens;

+++ app/Http/Controllers/ProjectController.php
@@ -61,7 +61,7 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
-        if($this->checkProjectOwner($id) == true){
+        if($this->checkPermissions($id) == true){
             return $this->service->update($request->all(), $id);
         }
         return response()->json(["error" => true, "message" => "Acesso negado!"], 412);
@@ -75,7 +75,7 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
-        if($this->checkProjectOwner($id) == true){
+        if($this->checkPermissions($id) == true){
             return $this->service->delete($id);
         }
         return response()->json(['error'=> true, 'message' => "Acesso negado!"]);
@@ -89,7 +89,10 @@ class ProjectController extends Controller
      */
     public function addMember(Request $request, $id)
     {
-        return $this->service->addMember($request->all(), $id);
+        if($this->checkPermissions($id) == true){
+            return $this->service->addMember($request->all(), $id);
+        }
+        return response()->json(['error'=> true, 'message' => "Acesso negado!"]);
     }
 
     /**
@@ -100,7 +103,10 @@ class ProjectController extends Controller
      */
     public function removeMember(Request $request, $id)
     {
-        return $this->service->removeMember($request->all(), $id);
+        if($this->checkPermissions($id) == true){
+            return $this->service->removeMember($request->all(), $id);
+        }
+        return response()->json(['error'=> true, 'message' => "Acesso negado!"]);
     }
 
     /**
@@ -110,7 +116,10 @@ class ProjectController extends Controller
      */
     public function members($id)
     {
-        return $this->service->members($id);
+        if($this->checkPermissions($id) == true){
+            return $this->service->members($id);
+        }
+        return response()->json(['error'=> true, 'message' => "Acesso negado!"]);
     }
 
     /**
